Split comm hooks into python-dependent hooks and others
